doc_analyzer_agent/requirements.txt
requests
beautifulsoup4
markdownify
google-generativeai
python-dotenv
playwright